% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars Dataset}
\format{
A data frame with 50 observations on 2 variables:
\describe{
\item{speed}{Speed of the car (in mph)}
\item{dist}{Stopping distance (in ft)}
}
}
\source{
Built-in R dataset
}
\usage{
cars
}
\description{
The \code{cars} dataset gives the speed of cars and the distances taken to stop.
}
\keyword{datasets}
