% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Calculate the Mean of a Numeric Vector}
\usage{
mymean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean value of the vector.
}
\description{
This function computes the mean of a vector.
}
